+++ src/App.jsx
@@ -1,15 +1,8 @@
 import React from 'react';
-// import { ProfCard } from './pages/prof-card/prof-card';
-import { CourseDescription } from './components/course-description/course-description';
 import './App.css';
 
 function App() {
-  return (
-    <div>
-      {/* <ProfCard /> */}
-      <CourseDescription />
-    </div>
-  );
+  return <div></div>;
 }
 
 export default App;
