+++ src/pages/CourseDescriptionPages/CoverCourseDescription/style.module.css
@@ -1,4 +1,4 @@
-.skill__card {
+.skillCard {
   width: 1160px;
   height: 310px;
   background-image: url(../../../../public/img/skill-card/skill-card.png);
