+++ src/pages/ProfCard/style.module.css
@@ -1,0 +1,72 @@
+.main {
+  width: 1160px;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 40px;
+  background: #271a58;
+}
+
+/* hover! */
+.block:nth-child(1):hover {
+  box-shadow: 10px -10px 16px purple, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
+  transition: 0.5s;
+}
+.block:nth-child(2):hover {
+  box-shadow: 10px -10px 16px blue, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
+  transition: 0.5s;
+}
+
+.block:nth-child(3):hover {
+  box-shadow: 10px -10px 16px orange, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
+  transition: 0.5s;
+}
+.block:nth-child(4):hover {
+  box-shadow: 10px -10px 16px greenyellow, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
+  transition: 0.5s;
+}
+.block:nth-child(5):hover {
+  box-shadow: 10px -10px 16px turquoise, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
+  transition: 0.5s;
+}
+
+/* class */
+
+.block {
+  display: block;
+  width: 360px;
+  height: 480px;
+  border-radius: 30px;
+  box-shadow: 10px -10px 16px rgba(0, 0, 0, 0.1), -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
+}
+
+.block:nth-child(1) {
+  background-size: 360px 480px;
+  background-image: url("../../../public/img/kurs-backgroundPicture/purple.png");
+}
+.block:nth-child(2) {
+  background-size: 360px 480px;
+  background-image: url("../../../public/img/kurs-backgroundPicture/blue.png");
+}
+.block:nth-child(3) {
+  background-size: 360px 480px;
+  background-image: url("../../../public/img/kurs-backgroundPicture/orange.png");
+}
+.block:nth-child(4) {
+  background-size: 360px 480px;
+  background-image: url("../../../public/img/kurs-backgroundPicture/green.png");
+}
+.block:nth-child(5) {
+  background-size: 360px 480px;
+  background-image: url("../../../public/img/kurs-backgroundPicture/blue-green.png");
+}
+
+.name {
+  font-family: "Abhaya Libre ExtraBold";
+  font-style: normal;
+  font-weight: 800;
+  font-size: 36px;
+  line-height: 110%;
+  padding-top: 30px;
+  padding-left: 30px;
+  color: #000000;
+}
