--- src/pages/prof-card/style.module.css
@@ -1,88 +1,0 @@
-.prof__card_main {
-  width: 1160px;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 40px;
-  background: #271a58;
-}
-
-/* hover! */
-.prof__card_block1:hover {
-  box-shadow: 10px -10px 16px purple, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-  transition: 0.5s;
-}
-.prof__card_block2:hover {
-  box-shadow: 10px -10px 16px blue, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-  transition: 0.5s;
-}
-
-.prof__card_block3:hover {
-  box-shadow: 10px -10px 16px orange, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-  transition: 0.5s;
-}
-.prof__card_block4:hover {
-  box-shadow: 10px -10px 16px greenyellow, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-  transition: 0.5s;
-}
-.prof__card_block5:hover {
-  box-shadow: 10px -10px 16px turquoise, -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-  transition: 0.5s;
-}
-
-/* class */
-.prof__card_block1 {
-  display: block;
-  background-size: 360px 480px;
-  background-image: url('../../../public/img/kurs-backgroundPicture/purple.png');
-  width: 360px;
-  height: 480px;
-  border-radius: 30px;
-  box-shadow: 10px -10px 16px rgba(0, 0, 0, 0.1), -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-}
-.prof__card_block2 {
-  display: block;
-  background-size: 360px 480px;
-  background-image: url('../../../public/img/kurs-backgroundPicture/blue.png');
-  width: 360px;
-  height: 480px;
-  border-radius: 30px;
-  box-shadow: 10px -10px 16px rgba(0, 0, 0, 0.1), -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-}
-.prof__card_block3 {
-  display: block;
-  background-size: 360px 480px;
-  background-image: url('../../../public/img/kurs-backgroundPicture/orange.png');
-  width: 360px;
-  height: 480px;
-  border-radius: 30px;
-  box-shadow: 10px -10px 16px rgba(0, 0, 0, 0.1), -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-}
-.prof__card_block4 {
-  display: block;
-  background-size: 360px 480px;
-  background-image: url('../../../public/img/kurs-backgroundPicture/green.png');
-  width: 360px;
-  height: 480px;
-  border-radius: 30px;
-  box-shadow: 10px -10px 16px rgba(0, 0, 0, 0.1), -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-}
-.prof__card_block5 {
-  display: block;
-  background-size: 360px 480px;
-  background-image: url('../../../public/img/kurs-backgroundPicture/blue-green.png');
-  width: 360px;
-  height: 480px;
-  border-radius: 30px;
-  box-shadow: 10px -10px 16px rgba(0, 0, 0, 0.1), -10px 10px 15.9988px rgba(0, 0, 0, 0.1);
-}
-
-.prof__card_name {
-  font-family: 'Abhaya Libre ExtraBold';
-  font-style: normal;
-  font-weight: 800;
-  font-size: 36px;
-  line-height: 110%;
-  padding-top: 30px;
-  padding-left: 30px;
-  color: #000000;
-}
